-Rebuilt architecture followed MVC model - Fix feature show doctors with role Patient.
