+++ doctorsCardiology.txt
@@ -1,4 +1,3 @@
 cdl7896| minh| ngoc| FEMALE| 12-05-2004| 20| Thach That, Ha Noi| +123-45678912| 12| 0| DOCTOR| CARDIOLOGY| 500.0
 cdl6789| nguyen trong| dat| MALE| 01-03-2000| 24| Thanh Mien, Hai Duong| +123-12345678| 12| 5| ASSOCIATE_PROFESSOR| CARDIOLOGY| 10000.0
 CDL1234| Minh| Nguyen| MALE| 07-09-2004| 19| Ha Noi| +123-12345678| 10| 10| DOCTOR| CARDIOLOGY| 23.0
-CDL3456| NGuyen| Huy| FEMALE| 6-9-2004| 19| Lang Son| +234-78675446| 10| 6| PROFESSOR| CARDIOLOGY| 12.0
