+++ src/Main.java
@@ -1,11 +1,11 @@
-import controller.ControllerMain;
+import viewLayer.ViewMain;
 
 import java.text.ParseException;
 
 public class Main {
     public static void main(String[] args) {
         try {
-            ControllerMain controller = new ControllerMain();
+            ViewMain controller = new ViewMain();
             controller.authenticationMenuTitle();
         } catch (ParseException e) {
             throw new RuntimeException(e);
